image_captioning_demo/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu
Pillow==7.0.0
matplotlib==3.2.2
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.4.1
streamlit==0.73.1
requests==2.25.1